가변인자 포인터 넘겨줌

+++ ft_printf/Libft/includes/libft.h
@@ -6,7 +6,7 @@
 /*   By: eunwolee <eunwolee@student.42seoul.kr>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/12/08 18:08:22 by eunwolee          #+#    #+#             */
-/*   Updated: 2022/12/17 15:09:07 by eunwolee         ###   ########.fr       */
+/*   Updated: 2022/12/17 15:57:37 by eunwolee         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -23,7 +23,7 @@ void	ft_striteri(char *s, void (*f)(char *));
 void	ft_toupper(char *c);
 size_t	ft_strlcpy(char *dst, const char *src, size_t dstsize);
 char	*ft_strchr(const char *s, int c);
-char	*ft_itoa_d(long arg);
+char	*ft_itoa_d(int arg);
 char	*ft_itoa_u(unsigned int arg);
 char	*ft_strdup(const char *s1);
